Move the variant to shared folder
(cherry picked from commit c26013e6a7884d4e4ceee6d0fb676bf4f11cc127)
