+++ ONI_Together/DebugTools/PacketTracker.cs
@@ -2,14 +2,12 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
-using System.Text;
-using System.Threading.Tasks;
 using ImGuiNET;
 using ONI_Together.Misc;
 using ONI_Together.Networking;
 using ONI_Together.Networking.Packets.Architecture;
+using Shared.OxySync;
 using Shared.Profiling;
-using Steamworks;
 using UnityEngine;
 
 namespace ONI_Together.DebugTools
