+++ ONI_Together/Networking/Components/LogicStateSyncer.cs
@@ -1,8 +1,6 @@
-using ONI_Together.DebugTools;
-using ONI_Together.Misc;
 using ONI_Together.Networking.Packets.World;
 using System.Collections.Generic;
-using Shared.Profiling;
+using Shared.OxySync;
 using UnityEngine;
 
 namespace ONI_Together.Networking.Components
