+++ ONI_Together/Networking/Packets/World/LogicStatePacket.cs
@@ -1,9 +1,8 @@
-using ONI_Together.Misc;
 using ONI_Together.Networking.Components;
 using ONI_Together.Networking.Packets.Architecture;
-using Shared.Interfaces.Networking;
 using System.Collections.Generic;
 using System.IO;
+using Shared.OxySync;
 using Shared.Profiling;
 using UnityEngine;
